[FIX] hr_payroll: prevent error when creating payslip with missing start date
Currently, an error is produced when the payslip start date is not set, but
the end date is provided.

**Steps to reproduce:**
- Install `hr_payroll` module(with demo).
- Create new **Payslip** for employee **Anita Oliver**.
- Clear both the start and end dates under the "Period" section.
- Set only the end date.

**Error:**
`TypeError: '<' not supported between instances of 'bool' and 'datetime.date'`

**Cause:**
The `_compute_warning_message` method performs a comparison between
`slip.date_from` and `contract_id.date_start` without verifying if `date_from`
is set. When `date_from` is `False`, the comparison `False < datetime.date(...)` - [1],
raises a `TypeError`.

[1] - https://github.com/odoo/enterprise/blob/a65314702b99df97c778d4f12f943191a19dafd1/hr_payroll/models/hr_payslip.py#L1029-L1031

This commit adds checks to ensure that start and end date is defined before performing
date comparisons in warning message computation, preventing the error.

Sentry - 6618275041

closes odoo/enterprise#87787

X-original-commit: b1d4b22378db2413289a192ef994465bcc768f35
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Aditi Patel (adip) <adip@odoo.com>