[IMP] ai_fields: Allow usage of binary fields in prompts
Purpose:
--------
Allow using files in AI prompts, such as "Make a summary of the {{pdf}}",
where `pdf` is a binary field or a many2one to `ir.attachment`.

To support this feature, the AI prompts are not rendered using QWeb anymore.
Instead, the fields inserted in the prompt are extracted, formatted, and
added inside a context dict sent along the user prompt (because one can not
insert the files inside the user prompt; they need to be sent separately to
the LLM).

This requires several changes:
- The fields are now stored in the prompt with the following format:
`<span data-ai-field="relation.field">Relation > Field</span>`
- The records are now stored in the prompt with the following format:
`<span data-ai-record-id="id">Display Name</span>`
- The usage of the `web_read` method to format the field values in the
prompt is replaced by a new `_ai_read` method

The fields and records are validated when a prompt is edited: fields must
be in the `mail_allowed_qweb_expressions` list, and read access is required
on the records.

The previous implementation using invisible elements and QWeb rendering
had several limitations: it required fields to be grouped within a single
multiline element, forced field values to be rendered inline in the prompt,
and often necessitated adding formatting logic directly within the prompt.

The new approach does the formatting in the backend based on the field type:
- Binary fields are extracted and sent in separate inputs as required by the
LLM APIs (as b64-encoded data for images and PDFs, or their indexed content
for other file types)
- Monetary fields are sent with their associated currency
- HTML fields will be stripped to their inner content only
- Relational fields include the display names of the records by default, but
the `_ai_format` method can be overridden on any model for which one needs to
send values other than the display names (e.g. `ir.attachment`)

Moreover, it enables building prompts better tailored for LLMs by eliminating
redundant information when a field appears multiple times and by adding field
values to a context dictionary (ORM snapshot) alongside the user’s prompt,
instead of replacing them inline in the user's prompt:
instead of sending to the LLM:
'Based on the {"name": ..., "partner_id": {name: ..., country: ...}, (...)}`
the user prompt will be of the form:
```
{
  "model_1": [{
    "id": 1,
    "partner_id": {"model": "model_2", "ids": 5}"
  }],
  "model_2": [{
    "id": 5,
    "name": ...,
    "country": ...,
  }]
}
```

Task-4762422

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>