[FIX] stock_barcode: udpate packaged lines on package scan
Steps to reproduce:

- In the settings: Enable `Units of Measure & Packagings`
- Create a storable product with 100 units in stock and an associated
  packaging such as 'pack of 6'
- Create and confirm a sale order for 5 pack of 6.
> This creates and reserve a delivery for 30 units (5 pack of 6).
- Process the delivery in the barcode app and scan a pack of 6
> A new line is created for 1 pack of 6 instead of updating the
> Quantity of the line referring to these packs to 5/30.

Cause of the issue:

Since the packaging/uom refactoring, packagings are just uoms. Now,
scanning the packaging will launch a call of the `_processBarcode`
method. During this call, the existing line will not be be found and
considered as an appropriate candidate to be updated since the its
prodcut_uom is in Unit and the packaging uom is not:
https://github.com/odoo/enterprise/blob/bdb3f3d959d6fd56e96e2299823fba6edbb5428c/stock_barcode/static/src/models/barcode_model.js#L1710-L1711
As such, there will be no `currentLine` to update and a new line will be
created:
https://github.com/odoo/enterprise/blob/bdb3f3d959d6fd56e96e2299823fba6edbb5428c/stock_barcode/static/src/models/barcode_model.js#L1441-L1443

opw-4782296

closes odoo/enterprise#87549

X-original-commit: 91cc5babaab608d2cc0d45f50236d2be89c4de82
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>