[FIX] knowledge: hide "invite" button for portal users
Currently, the "Invite" button is visible to portal users. When a portal
user clicks this button, the system attempts to open a wizard using the
`knowledge.invite` model. However, because portal users lack access
rights to this model, the system raises an "Access Error".

Steps to reproduce:
1. Log in as a portal user
2. Open an article
3. Click on the "Share" button
4. Click on the "Invite" button

=> Traceback: Access Error on the model `knowledge.invite`

To fix the issue, we will simply hide the "Invite" button for the portal
user as it was the case before the permission panel refactoring.

See: odoo/enterprise#76261

Task-4636494

closes odoo/enterprise#87207

X-original-commit: f82525455a0c48cb5cd616f79bc235d7b4dca84e
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>