[IMP] iot: show dns error message
When the user pairs an IoT box in the SaaS, it automatically receives
a TLS certificate from odoo.com so that it can be accessed via HTTPS.
However, sometimes the user's DNS will not be able to resolve the IoT
box's new address straight away, making it unreachable.

We cannot detect this error directly, and there is nothing we can do on
our end to alter the user's DNS setup. Instead, this commit adds an
error message mentioning the DNS issue in the event that an IoT box has
a valid certificate but we are unable to reach it.

task-4787494

closes odoo/enterprise#86371

X-original-commit: 7147a31e8eaae6d9f1e168ab9dfbd310db99d0ce
Related: odoo/odoo#211573
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>