[FIX] account_reports: currency rate input
Update on the filterExchangeRate function.

The text input allowed users to enter non-numeric characters, which led to
float conversion errors during processing.

Now we parse the string into a float, catch errors and notify the user that the number was not valid.
It also enables the user to use the thousands and decimal separators based on the language of the user.

opw-4788246

closes odoo/enterprise#86259

X-original-commit: e6b5d0f5ae80720605fb9224da7f8145754fde5e
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>