[FIX] documents: fix m2o conversion to object
Purpose
=======
Fix the parent folder placeholder which is set to "Shared with me"
instead of "My Drive" when creating a folder in My Drive.
Fix the traceback appearing when trying to share a shortcut.
Fix the PDF splitted parts which aren't saved in the save folder as
the original PDF.

Specification
=============
Following the PR linked, in the RelationalModel the value of
a many2one is now represented using an object instead of an array.

While converting the use in the documents module, some occurrences
have been missed resulting in strange front-end behavior.
Fixing that by making sure all the many2one values are correctly
accessed using an object key.

related PR: odoo/enterprise#83203

Task-4762955

closes odoo/enterprise#84532

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>