[FIX] sign: redirect using request item
Before this commit:
When closing the popup after signing an document, it redirects to the
`/my` page instead of `/my/signature/<id>`. It only happens when the
request's id is out-of-sync with the request item's id.

After this commit:
Correctly use the `sign.request.item`'s id, as expected from the
`/my/signature` route.

closes odoo/enterprise#86690

Signed-off-by: Florian Vranckx (flvr) <flvr@odoo.com>