[IMP] account_accountant: allow filling with reco model on partial
So, with the idea of directly reconciling the aml and leave the
partial on the statement line if the amount is close, we want an easy
 way for a user to assign these leftovers.

Before this commit, when you clicked on the reconciliation model after
partially putting amount, it would try to override everything on the
statement line and putting what is said by the model.
Now, we leave the amounts previously reconciled and complete it with
the model.

task-4749343

X-original-commit: 34fc3fdef30ad89128653e66d92137065364503c
Part-of: odoo/enterprise#87224
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>