[FIX] l10n_ca_payment_cpa005: stop relying on One2many ordering
The order of a One2many field isn't well defined and depending on the
situation the records may be in the order you added them or the order
defined by `_order` on the model. Before 18.3, a hack was added to
sort them using `id` which worked. In 18.3 it no longer works
consistently.

This explicitly orders the payments when generating the file using
their default _order, which is the default order in which they appear
in the batch payment list view.

Fixes runbot build error 181847

closes odoo/enterprise#86899

X-original-commit: 2934af8bc6fcba45cad51be00dbbda401f38227c
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>