[FIX] l10n_co_dian: avoid error when sending invoice
Currently, sending an invoice can raise an error if the
**Electronic Invoice Type** is not set on the invoice.

**Steps to Reproduce:**
- Install the `l10n_co_dian` module.
- Change the default company to **CO Company**.
- Create and confirm a customer invoice, but leave the **Electronic Invoice Type** unset.
- Send the invoice.

**Error:**
`AttributeError - 'bool' object has no attribute 'rjust'`

Here, an issue occurs because the system attempts to call `rjust()` on the
field `l10n_co_edi_type` without ensuring it is a string. When the field
is `False`, it results in an Attribute error. - [1]

This commit ensures the right access to `rjust()` only when
`l10n_co_edi_type` is a valid string, preventing the error.

[1] - https://github.com/odoo/enterprise/blob/8d7a13822e9fd480a6e279ff93e29834a6668536/l10n_co_dian/models/account_edi_xml_ubl_dian.py#L974-L975

Sentry - 6622171121

closes odoo/enterprise#86893

X-original-commit: 95cb03ff9a00dd22b93a0bef715779f3a4e12ff2
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Aditi Patel (adip) <adip@odoo.com>