[REF] iot: remove ip_url computed field
The `ip_url` field on the IoT box model was intended to provide the
full URL of the IoT box, e.g. 'http://1.2.3.4:8069' or
'https://1-2-3-4.abcd.odoo-iot.com'. However the problem was that it
relied on `web.base.url` to determine if it should use HTTP or HTTPS,
which can cause problems if the Odoo database can be accessed via both.

To fix this, we instead started to use a helper on the frontend,
`formatEndpoint`, to perform the same task, except that it can take into
account the actual browsing context and always use the same protocol.

This changes the remaining uses of `ip_url` to use `formatEndpoint`,
except the `open_homepage` action which now gets the current protocol
from the request object. With no remaining usages the field itself can
be removed.

task-4813536

closes odoo/enterprise#86273

Related: odoo/upgrade#7735
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>