[FIX] documents,website_documents: fix fields visibility logic
Fixes the visibility of the `website_id` and `company_id` fields.

Previously, in commit ddf102e5, we addressed this differently
but it turns out that there's no simpler way to handle it in
master, so we apply this change in stable as well.

These fields should be hidden from non-internal users AND
**only** when multiple websites or multiple companies are
available. Templates do not allow to AND groups, so we
revert to "multi-" group there and check group_user in JS.

Includes HOOT tests to check the visibility of `company_id`
for portal and internal user.

Task-4505668

closes odoo/enterprise#91005

X-original-commit: 91031b147997fa1d330b55dc076af1d08570d3ec
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>