[FIX] l10n_mx_edi{_pos}: raise user error outside localization
Currently, there is an button related to the MX localization that
can be accessed outside the localization if `l10n_mx` is installed.

Steps to reproduce:
-------------------
* Install `l10n_mx`
* With company other than MX, go to the pos orders
* Select multiple pos orders, select the button **Actions**
> Observation: The action `Create Global Invoice` is visible
* Select the action `Create GLobal Invoice`
> Validation error, Selected orders are not eligible for CFDI

Why the fix:
------------
This is an `ir.actions.server` which is introduced in the `l10n_mx`
localization.

It does not look possible to add `invisible` on the xml for the action.
https://github.com/odoo/odoo/blob/2540155b5e00a6547fcceee4af4b83f385dcd4ad/odoo/addons/base/models/ir_actions.py#L179-L215

What we can do instead is to throw a more explicit error when trying
to use the action outside the MX loca.

opw-4805257

closes odoo/enterprise#87233

X-original-commit: fdf764c1fe84c972189dd14fdd6a6ccb0ed0ea09
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>