[FIX] web_studio: fix webhook automations export
Before this commit, the base.automation.url field is exported but it is
a computed field that depends on the webhook_uuid field, which is not
exported. Later when the generated export is imported, the url field is
computed from a new webhook_uuid. So the exported url is not really
useful.

More importantly, the url/webhook_uuid is a private information that
should not be exported.

After this commit, the url field is no more exported.

X-original-commit: f92023d7ca97526706c68817d4a877731015ba44
Part-of: odoo/enterprise#86753
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>