[FIX] voip: send keep-alives to avoid silent disconnection
PR #64267 implemented a reconnection mechanism, but it is only triggered
when the WebSocket close event is fired. In practice, we noticed that
the WebSocket would sometimes silently lose connection. When this
happens, one should just send data through the WebSocket to refresh it.

This commit enables SIP.js keep-alive feature, hopefully reducing the
number of disconnections, or at least preventing them to go unnoticed.

closes odoo/enterprise#86365

X-original-commit: 26b8521b45b6c6abbcf0e35fd47948e27f7cd868
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>