[IMP] knowledge: fix editor padding to enhance editing experience
With version 18.1, we migrated from legacy HTML editor to the new one.
As part of this upgrade, we revised the templates and reworked the wysiwyg
helper. During the migration, several CSS classes were altered or removed.

One issue that emerged is that the CSS rules intended to add bottom
padding to the editor area are still present in the stylesheet but no
longer apply to any element. These rules were especially helpful when
working with long-form content, as they allowed users to scroll further
and bring the end of the text closer to the center of the screen -
improving readability and enhancing the editing experience.

To fix this, we'll update the relevant CSS selector to ensure the padding
is correctly applied to the editable area.

See: odoo/enterprise#67083

Task-4636494

closes odoo/enterprise#91221

X-original-commit: bc1e65552847dcfccb55932d675b059b2f63d2e0
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>