[IMP] *external_tax,*avatax*: consolidate and simplify code
Before this, the modules made heavy use of small methods that needed
to be overridden to adapt the behavior. These methods could be on
various models, e.g.,:

- account.external.tax.mixin: to provide hooks that need to be
overriden per model (e.g. account.move, sale.order, ...)
- account.move/sale.order: to provide hooks for just one record

And in different modules, e.g.,:

- account_external_tax: if the behavior is generic enough for all
external tax modules,
- account_avatax/l10n_br_avatax: for behavior only required in certain
modules,

The combination of this made the flow of execution hard to follow and
the code hard to understand. This simplifies things by reducing the
amount of methods that need to be overridden.

To achieve that each tax integration defines a method that gathers
"service parameters". This is a dictionary per record that contains
all data needed to build the API request. This is the primary method
to override to fill out model-specific values (e.g. what is the tax
date for a sale.order). It now uses accounting base lines to represent
the data of lines (account.move.line, sale.order.line, etc), instead
of doing something ad-hoc like before.

The methods that follow in the flow exclusively use these service
parameters and can therefore be @api.model methods. The payload for
the API is prepared in "prepare document" methods. The API call is
made and tax results are processed in an "extract tax values"
method. The extracted tax values are then used to create missing taxes
and tax groups in _process_external_taxes and finally used to set the
right manual_tax_amounts in _set_external_taxes.

As a sidenote, the dedicated "commit" endpoint in the Avalara API was
implemented for account_avatax to commit transactions (mark them as
final so Avalara includes them in tax filings). Before this, a whole
new tax calculation transaction was created with "commit" set. With
this new approach, it was much simpler to implement the commit
endpoint to achieve the same thing.

task-4601840

closes odoo/enterprise#82623

Signed-off-by: Laurent Smet (las) <las@odoo.com>