[IMP] {account,l10n_be}_disallowed_expenses: rate date and display
We have included deductible expense rates for Belgium - but the starting dates default to the module installation date.
In most cases they should apply historically too.
Since we don't have the exact start dates, we will default to the beginning of 2025.

When there is no rate defined, it means 100% deductible, however the UI displays nothing instead of 100%.
Default to 100% in all views, including the report.

task-4813998

closes odoo/enterprise#86565

X-original-commit: 681bdb283c582954c11eb470c3d3fcf628e57009
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>