[IMP] {l10n_us,}hr_payroll_account: write payment date on payment report creation
When creating a payment report, the user must can choose the effective date. The term is ambiguous and is replaced by this commit by payment date. Additonnaly, the pay_date field on the payslip is automatically filled according to the payment date of the attached payment report.

task-4814800

Part-of: odoo/enterprise#86274
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>