[FIX] spreadsheet_dashboard_hr_contract: fix invalid domain
Alright, ready? Story time.

Commit e33522cdc8fcc attempted to fix some domains.
But it was wrong. I spotted the mistake and patched it during the forward-port
to saas-18.3 (e30f0e69820) (*)

So far, so good? Not really.
Turns out that patch was also broken because there
an extra invalid '&'

This commit fixes this issue.

Holy moly, where was my brain that day?!

*it was fixed in 18.0,18.1 and 18.2 later with 27e1696d9e414

X-original-commit: 2709e3bcf7fe740b1882fae2590b3bff2bc5d00b
Part-of: odoo/enterprise#87006
Related: odoo/odoo#212972
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>