[FIX] account_batch_payment: Adapt test and pyaments domains
This commit updates the domain to include all valid payment states as
returned by `_valid_payment_states()`.

test_validate_batch was failing with single app installed when accountant module was
not installed. When only Invoicing, we allow validating batch with paid
payments, so the Warning was not raised causing tests to fail.

Apapt the test to handle the case when accountnat is not installed.

closes odoo/enterprise#86483

X-original-commit: 95567f106fba3ed6d334ce6ad523378239865204
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Seif Gneedy (segn) <segn@odoo.com>