[IMP] mrp: move description_bom_line to description_picking
The computed field `description_bom_line` on `stock.move` shows whether
these moves are part of a kit.
Instead of making this a separate field that needs to be stored on the
moves, we now put this information into the `description_picking` field.

task-4783553

closes odoo/enterprise#86304

Related: odoo/odoo#211488
Related: odoo/upgrade#7736
Signed-off-by: Claeys Pieter (clpi) <clpi@odoo.com>