[IMP] account_reports: Including account_return in lock dates
Before the closing move was not included in the tax lock date and could
be reset while being locked. Now we also lock the closing move.

Same for the return, we lock the return if the period is locked. Before,
when we wanted to reset the return, it would automatically revert the
lock date. Now user should do it manually otherwise he will not be able
to reset the return.

task-4818752

closes odoo/enterprise#91276

X-original-commit: 98abe4fbfb9938c06fc8e9ffc28b0db951404db0
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>