[FIX] l10n_fr_fec_import: Fixed unit test file reading
The JS unit test of the FEC parser was having a flaky build error because
for reading files, the test was using `.getReader().read()`
which would read only one chunk from the stream, causing incomplete files to be processed for large files.
This commit makes the unit test use `response.arrayBuffer()` instead
to guarantee the processing of the complete files all the time.

build_error-161014

closes odoo/enterprise#86453

X-original-commit: f6b94abf1a4cb71b93e2e5522f8e1f4e34fb2c0d
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>