[FIX] mail_enterprise: livechat visitor counter badge is red
This was green before this commit, matching the community badge
color. Enterprise is red, which is the prettiest and most well-known
color for badge, but due to being unavailable in bundle it was
shown as green rather than red.

Discuss public page also lacked it, which this commit fixes too.

closes odoo/enterprise#86456

X-original-commit: 4acd3a719378093f6e73286f30f5d94edd8c9f9a
Related: odoo/odoo#211730
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>