[FIX] sale_timesheet_enterprise: prevent error when opening tasks in portal
When Invoicing Policy is set to Validated timesheets only, opening tasks in
the portal view results in a traceback.

Steps to reproduce the error:
- Install "sale_timesheet_enterprise"
- Go to Timesheets > Configuration > Settings >
  Invoicing Policy: Validated timesheets only
- Switch to Portal View > Open Tasks

Traceback:
ValueError: Aggregate method is mandatory for 'unit_amount'

https://github.com/odoo/enterprise/blob/3cfd66bfaf6a1fd58200b46b0c56330b8d96dcdd/sale_timesheet_enterprise/models/project_task.py#L130
Here, Aggregate method is missing for the "unit_amount" field.
So, It will lead to the above traceback.

sentry-6626038923

closes odoo/enterprise#86601

X-original-commit: c12ef0ebe32d585b6298a0ec37877489b841090b
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Renilkumar Kajavadra (reka) <reka@odoo.com>