[FIX] l10n_mx_edi_website_sale: ensure test data is set up
The test relies on the admin partner having full contact in order for
the "Confirm Address" step to "just work". If the test is run without
demo data, this is rather not the case and the test fails on the next
step, as "Confirm Address" causes a form submission failure due to
required fields not being filled.

https://runbot.odoo.com/odoo/error/161633

closes odoo/enterprise#86421

X-original-commit: 3f866eed25fd7cccb5d85da86a8c2ae1144576f8
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>