[FIX] helpdesk_timesheet: fix rounding of total hours spent
to reproduce:
=============
- enable timesheet for a helpdesk team
- create 5 tickets with 10 minutes each
- on list view of tickets, select these tickets

-> total spent time is 51 minutes instead of 50 minutes

Problem:
========
time is represented as float, when computing the total time spent, we round
the float value to 2 decimal places, which can lead to loss of precision.
for example 10 minutes is represented as 0.16666666666666666 hours, and when
we round it to 2 decimal places, we get 0.17 hours, which is 10.2 minutes.
when we sum up 5 tickets, we get 0.85 hours, which is 51 minutes.

Solution:
=========
we don't round the float value to 2 decimal places, and we let the front
handle the display of the time in the correct format.

opw-4560872

closes odoo/enterprise#87061

X-original-commit: 7d4fca3739443abbdca85e82dccb892d1bb138e3
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>