[FIX] sale_commission: ensure the achievement id is based on cl.date
Before this commit, when several achievements were done on the same
period for the same record, they would share the same id and therefore
could not be properly processed in the interface.

In order to fix this, we decide to use the achievement date to compute
the id. As a result, for achievement based on MRR (subscription), if two
sale.order.log corresponds to the same order, on the same date, we will
end up with a duplicated id too.
It happens because the related_res_model and related_res_id of
achievement based on MRR is the order linked to the sale.order.log.

To avoid this issue, we introduce another source of entropy in the
computation of the achievement's id: the write date of the record.
In that case, we are free to use the write_date of the sale.order.log
and therefore they should be different.
Same apply for sale.order and account.move for other kind of
achievements because of the UNION ALL.

closes odoo/enterprise#86644

closes odoo/enterprise#87097

Taskid: 4845096
X-original-commit: 5b6fdda126e4cfa5ebf92f7b96d1805f6c4b992b
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>