[REF] sale_subscription: simplify catalog logic
Now that the recurring prices are correctly handled through the
pricelist inner logic, we only have to forward the plan so that
it's considered to compute the prices.

Also moves the logic where it belongs as overrides of the mixin
only for `sale.order` records, instead of overrides of the generic
controller routes.

task-3830130

Part-of: odoo/enterprise#84396
Related: odoo/odoo#207912
Related: odoo/upgrade#7698
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>