[FIX] knowledge: fix z-index to display sidebar above chatter topbar
When the user opens the chatter and expands the sidebar, the chatter's
topbar overlaps the sidebar. This overlap disrupts the layout and can
be frustrating, as it may hide some article links. To fix this, we'll
update the sidebar's z-index so it sits above regular elements but
remains below modals.

Task-4636494

closes odoo/enterprise#87082

X-original-commit: 805e55d1c41c37dc9b0dbbae26de6e020eede5b0
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>