[FIX] mrp_workorder: close edit operators
How to reproduce
================
1. Open Shop Floor;
2. Click on "Edit Operators";
3. Close the "Edit Operators" by clicking on the close button (x on the
   top right) => Traceback 💥

Issue
=====
When the diablog is closed, the `MrpDisplay` `reload` method is called.
This method can have a parameter, `record` which is not defined (and
thus `false` by default) when not needed, but when a dialog is closed by
the close button, a special dict is given as the parameter, and this
object is used as the `record` in `reload` which is wrong.

Fix
===

Check if `record` has no `dismiss` key before assuming `record` is
actually a record.

closes odoo/enterprise#86650

closes odoo/enterprise#87611

Related: odoo/odoo#212157
Related: odoo/odoo#214181
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>