[IMP] account_accountant: link to the bank
Before this commit, when adding a line that is not from a move we would have a
link to the bank. User should not be able to go to the bank from there to avoid
mistakes where they will change the move. Specially when the move is fully reconciled.
If the move is partially reconcile they have access to the move in the dropdown,
if it's really needed. They also have all the information in the chatter.

no task id

closes odoo/enterprise#86938

X-original-commit: 9043fc314be884968f6f946af3b111464b07e897
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>