[FIX] l10n_mx_edi_pos: use correct issued adress order global invoice
Steps to reproduce:
[l10n_edi_extended]
- setup up a mexican company
- create a journal in which you define an issued address (different from your company)
- setup your pos with newly created journal
- Open a pos session
- add a product
- pay in cash with no customer
- in the backend go into the orders
- select the newly created order in the list view and click on the action "create global invoice"
- in the order > cfdi: download the created xml

Issue:
The "Lugar de expedicion" will have the zip code of the company and not from the issued address one

Cause:
We only check for issue adress in account_move but not from orders. Global invoice is not only for invoices but also to account for sale (orders) made with petty cash

Solution:
We don't want to create a bridge module only for that. We check if the field is defined on `account.journal`

opw-4802389

closes odoo/enterprise#90712

X-original-commit: 3bddb94b8df6c46d8e69868aae7ec4393c3b03c2
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>