[FIX] industry_fsm: respect order defined by user to find default project
Before this commit, when the user creates more than one fsm project,
the default fsm project selected once a fsm task is created if the one
with the lowest sequence and lowest id which could be confusing for the
user since the default order displayed in the list view of projects is
`sequence, name, id`.

This commit makes sure the default fsm project fetched is the first one
displayed in the list view of projects.

opw-4791468

closes odoo/enterprise#86856

X-original-commit: e1dc29d120dde304166fde55d946998d062dedf5
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>