[IMP] account_accountant: automatic model for fees
For the leftover from the full reco of the invoice that
are left on the statement line, we propose a new reconciliation
model that will be put when we encounter these situations.

We create it when the user is in the case of the 3% and that
he put the leftover on a specific account.

task-4749343

X-original-commit: 7d611b67555882b4bd737c1d811006b9e7df0683
Part-of: odoo/enterprise#87224
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>