[FIX] l10n_sg_reports: fix initial_balances in iras audit export
Before this commit the opening balance in the iras audit export was always at 0.

This was caused because it would call _query_values which was not
returning the initial balances.

The fix was to instead call _get_initial_balance_values. With this we
could finally get the opening balances.

task-4826744

closes odoo/enterprise#87144

X-original-commit: d29b82daf28bbacdfe52a7ee917158139263e498
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>