[FIX] pos_blackbox_be: 2 traceback opening restaurant
This commit fixes two traceback that happen when
someone opens a pos config with a belgian blackbox configured.
The first traceback happens because of the call of navigate throught
this.pos while the call is made inside of pos_store. The other traceback
is cause by the _notify "CLOCKING" where the "_" fields are not sent.

closes odoo/enterprise#86605

Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>