[FIX] ai: dropdown items over modals
Steps to reproduce:
- Install any module with exposed views (say sale_management) and the ai module.
- Open any list view (say the Quotations) and type something in the search bar.
- Dropdown menu items will appear and click "Custom Filter" button from the items.
- [ISSUE] The items are shown over the custom filter dialog.

The issue is caused by the custom rules introduced in the ai module. In this
commit, we're removing all the custom rules that overrides the overlay items. We
shouldn't be controlling the `z-index` of those items.

To ensure that the AI ChatWindow is still above the dialog that opens it, we
reduce the `z-index` of the dialog to `$zindex-sticky=1020` (from
`zindex-modal=1055`) and change the `z-index` of the ChatWindows to
`$zindex-sticky + 1`. This works because the `.o-mail-ChatWindow` elements lives
in the same "stacking context" as the `.o-overlay-item` that contains the
dialog. In the dom (and in terms of stacking context) it roughly looks like the
following:

- html (root stacking context)
  - .o-overlay-item:has(.o_dialog) (z-index=1020) (under .o-overlay-container)
  - .o-mail-ChatWindow (z-index=1021) (under .o-mail-ChatHub)
  - .o-mail-ChatHub-bubbles (z-index=1020)

With the above `z-index` values, the stacking will be:
- .o-overlay-item:has(.o_dialog)
- .o-mail-ChatHub-bubbles
- .o-mail-ChatWindow (top most)

The bubbles are not visible when the chat window is open, and when the
chatwindow is minimized, the bubbles is still visible because it's above the
dialog.

closes odoo/enterprise#86734

X-original-commit: 43a82f082eb4d658281b40b1bba8b044da4567a8
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>