[FIX] documents: fix `_search_last_access_date_group` when False is in values.
A small error was been made in
https://github.com/odoo/enterprise/pull/76079
(`_search_last_access_date_group`): the SQL string contains a %s
without an argument, resulting in a "TypeError: not enough arguments
for format string."

Fix it. A test will be added later to the master branch in a later PR
that found the issue https://github.com/odoo/odoo/pull/133224.

closes odoo/enterprise#86819

X-original-commit: 8fb7e0530b546d3ec1886ff0ee64c1c326661649
Related: odoo/odoo#212605
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>